requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pygame==2.6.1
PyJWT==2.10.1
sqlparse==0.5.3
